Fixed some default CVAR values
Unlocked and raised `r_lodCurveError` value as it is in default urt client.
FrozenSand/ioq3-for-UrbanTerror-4@e0023c1

Increase default `cl_maxpackets` to `125`
This decreases latency if FPS is greater than or equal to 60
(since the previous default for `cl_maxpackets` was `60`), and
allows for more consistent input handling.

Increase default geometric detail on characters and patches
`r_subdivisions`
`r_lodbias`